requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
defusedxml==0.7.1
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
httpie==3.2.1
idna==3.4
multidict==6.0.3
Pillow==9.2.0
pycodestyle==2.9.1
Pygments==2.13.0
PyJWT==2.6.0
PySocks==1.7.1
pytz==2022.2.1
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.2
urllib3==1.26.13
